// src/features/radicacionCasos/types.ts
export interface NuevoAfiliadoForm {
    tipoId: string
    id: string
    nombres: string
    apellido1: string
    apellido2: string
    sexo: string
    direccion: string
    telefono: string
    fechaNacimiento: string
    municipioCodigo: string
    municipioNombre: string
    departamento: string
    regimen: string
    tipoCotizante: string
    eps: string
}